move more code around
