+++ unison-core/unison-core1.cabal
@@ -32,6 +32,7 @@ library
       Unison.DataDeclaration.ConstructorId
       Unison.DataDeclaration.Names
       Unison.DataDeclaration.Records
+      Unison.DeclNameLookup
       Unison.Hashable
       Unison.HashQualified
       Unison.HashQualifiedPrime
